10. add FSD structure
