+++ src/app/providers/themeProvider/index.tsx
@@ -1,0 +1,5 @@
+import { ThemeContext } from "./lib/ThemeContext";
+import { useTheme } from "./lib/useTheme";
+import { ThemeProvider } from "./ui/ThemeProvider";
+
+export {ThemeContext, ThemeProvider, useTheme}
