+++ src/app/types/global.d.ts
