--- src/greet.ts
@@ -1,3 +1,0 @@
-export const greet = (name: string) => {
-  console.log(`Hello ${name}`);
-};
