+++ src/index.tsx
@@ -1,7 +1,7 @@
 import { render } from "react-dom";
-import App from "./App";
+import App from "app/App";
 import { BrowserRouter } from "react-router-dom";
-import { ThemeProvider } from "./themes/ThemeProvider";
+import { ThemeProvider } from "app/providers/themeProvider/ui/ThemeProvider";
 
 render(
   <ThemeProvider>
