+++ src/pages/AboutPage/index.tsx
@@ -1,0 +1,1 @@
+export {AboutPageAsync as AboutPage} from './ui/AboutPage.async'
